Fix menu item price validation to accept float values from Deliverect
- Changed price from int to float in MenuItemBase and MenuItemUpdate
- Deliverect sends decimal prices like 2.5, not integer cents
- This resolves the second Pydantic validation error blocking menu updates
- Modifier groups stored successfully, now items should work too

+++ app/schemas/menu.py
@@ -118,7 +118,7 @@ class MenuCategoryUpdate(BaseModel):
 class MenuItemBase(BaseModel):
     name: str
     description: Optional[str] = None
-    price: int  # Price in cents  
+    price: float  # Price (can be decimal)  
     plu: Optional[str] = None  # Changed to optional to match model
     category_id: Optional[int] = None  # Changed to optional to match model
     deliverect_item_id: Optional[str] = None  # Added missing field
@@ -138,7 +138,7 @@ class MenuItemCreate(MenuItemBase):
 class MenuItemUpdate(BaseModel):
     name: Optional[str] = None
     description: Optional[str] = None
-    price: Optional[int] = None
+    price: Optional[float] = None
     plu: Optional[str] = None
     category_id: Optional[int] = None
     deliverect_item_id: Optional[str] = None  # Added missing field
